requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
dotmap==1.3.30
executing==2.2.1
fastapi==0.127.1
fonttools==4.60.1
git-filter-repo==2.47.0
greenlet==3.3.0
h11==0.16.0
heartpy==1.2.7
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.7.0
jupyter-core==5.9.1
kiwisolver==1.4.9
lazy-loader==0.4
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mne==1.11.0
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.1
pooch==1.8.2
prompt-toolkit==3.0.52
psutil==7.1.3
pure-eval==0.2.3
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyparsing==3.2.5
pyppg==1.0.14
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sqlalchemy==2.0.45
stack-data==0.6.3
starlette==0.50.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.6.2
uvicorn==0.40.0
wcwidth==0.2.14
